Move test of building with envopt to "env_opts_test".
