+++ ailets-rs/messages_to_query/tests/structure_builder_test.rs
@@ -248,138 +248,3 @@ fn support_special_chars_and_unicode() {
     );
     assert_that!(writer.get_output(), equal_to(expected));
 }
-
-fn _build_with_env_opts(env_opts: EnvOpts) -> String {
-    let writer = RcWriter::new();
-    let mut builder = StructureBuilder::new(writer.clone(), env_opts);
-    builder.begin_message().unwrap();
-    builder.add_role("user").unwrap();
-    builder.begin_content().unwrap();
-    builder.begin_content_item().unwrap();
-    builder.begin_text().unwrap();
-    write!(builder.get_writer(), "Hello!").unwrap();
-    builder.end_text().unwrap();
-    builder.end_content_item().unwrap();
-    builder.end_content().unwrap();
-    builder.end_message().unwrap();
-    builder.end().unwrap();
-    writer.get_output()
-}
-
-#[test]
-fn override_endpoint_model_stream() {
-    let mut opts = HashMap::new();
-    opts.insert(
-        "http.url".to_string(),
-        serde_json::Value::String(
-            "https://my-custom-fairy-api.example.com/v1/chat/completions".to_string(),
-        ),
-    );
-    opts.insert(
-        "llm.model".to_string(),
-        serde_json::Value::String("my-custom-fairy-model".to_string()),
-    );
-    opts.insert("llm.stream".to_string(), serde_json::Value::Bool(false));
-    let env_opts = EnvOpts::from_map(opts);
-
-    let output = _build_with_env_opts(env_opts);
-
-    assert_that!(
-        output.as_str(),
-        matches_regex("my-custom-fairy-api.example.com")
-    );
-    assert_that!(output.as_str(), matches_regex("my-custom-fairy-model"));
-    assert_that!(output.as_str(), matches_regex("\"stream\": false"));
-}
-
-#[test]
-fn override_stream_option() {
-    // set to "false"
-    let mut opts = HashMap::new();
-    opts.insert("llm.stream".to_string(), serde_json::Value::Bool(false));
-    let env_opts = EnvOpts::from_map(opts);
-    // act and assert
-    let output = _build_with_env_opts(env_opts);
-    assert_that!(output.as_str(), matches_regex("\"stream\": false"));
-
-    // set to "true"
-    let mut opts = HashMap::new();
-    opts.insert("llm.stream".to_string(), serde_json::Value::Bool(true));
-    let env_opts = EnvOpts::from_map(opts);
-    // act and assert
-    let output = _build_with_env_opts(env_opts);
-    assert_that!(output.as_str(), matches_regex("\"stream\": true"));
-
-    // set to an invalid value
-    let mut opts = HashMap::new();
-    opts.insert(
-        "llm.stream".to_string(),
-        serde_json::Value::String("invalid".to_string()),
-    );
-    let env_opts = EnvOpts::from_map(opts);
-    // act and assert
-    let output = _build_with_env_opts(env_opts);
-    assert_that!(output.as_str(), matches_regex("\"stream\": true"));
-}
-
-#[test]
-fn add_llm_options_of_different_types() {
-    // Test a string value
-    let mut opts = HashMap::new();
-    opts.insert(
-        "llm.foo".to_string(),
-        serde_json::Value::String("bar".to_string()),
-    );
-    let env_opts = EnvOpts::from_map(opts);
-    let output = _build_with_env_opts(env_opts);
-    assert_that!(output.as_str(), matches_regex("\"foo\": \"bar\""));
-
-    // Test a number value
-    let mut opts = HashMap::new();
-    opts.insert(
-        "llm.max_tokens".to_string(),
-        serde_json::Value::Number(serde_json::Number::from(100)),
-    );
-    let env_opts = EnvOpts::from_map(opts);
-    let output = _build_with_env_opts(env_opts);
-    assert_that!(output.as_str(), matches_regex("\"max_tokens\": 100"));
-
-    // Test an array value
-    let mut opts = HashMap::new();
-    let arr = vec![
-        serde_json::Value::String("system".to_string()),
-        serde_json::Value::String("user".to_string()),
-    ];
-    opts.insert(
-        "llm.allowed_roles".to_string(),
-        serde_json::Value::Array(arr),
-    );
-    let env_opts = EnvOpts::from_map(opts);
-    let output = _build_with_env_opts(env_opts);
-    assert_that!(
-        output.as_str(),
-        matches_regex(r#""allowed_roles":\s*\["system","user"\]"#)
-    );
-
-    // Test an object value
-    let mut opts = HashMap::new();
-    let mut obj = serde_json::Map::new();
-    obj.insert(
-        "top_p".to_string(),
-        serde_json::Value::Number(serde_json::Number::from_f64(0.9).unwrap()),
-    );
-    obj.insert(
-        "top_k".to_string(),
-        serde_json::Value::Number(serde_json::Number::from(50)),
-    );
-    opts.insert(
-        "llm.sampling_params".to_string(),
-        serde_json::Value::Object(obj),
-    );
-    let env_opts = EnvOpts::from_map(opts);
-    let output = _build_with_env_opts(env_opts);
-    assert_that!(
-        output.as_str(),
-        matches_regex(r#""sampling_params":\s*\{"top_k":50,"top_p":0.9\}"#)
-    );
-}
